pyproject.toml
[project]
name = "projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-openai>=0.3.32",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
    "tiktoken>=0.11.0",
    "youtube-transcript-api>=1.2.2",
]
